Moved the enqueue_worker method to a helpers module. Moved the EnqueueClient to the sockets module. Removed the queue module.
