+++ bld_server/src/sockets/mod.rs
@@ -1,8 +1,10 @@
+mod enqueue_client;
 mod exec;
 mod exec_client;
 mod ha;
 mod monit;
 
+pub use enqueue_client::*;
 pub use exec::*;
 pub use exec_client::*;
 pub use ha::*;
